Add proximity sensor interface files
